Split send signal and sequence into own contribs
