--- src/vs/workbench/contrib/terminalContrib/sendCommands/common/terminal.sendCommands.ts
@@ -1,9 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-export const enum TerminalSendCommandsCommandId {
-	SendSequence = 'workbench.action.terminal.sendSequence',
-	SendSignal = 'workbench.action.terminal.sendSignal',
-}
